watcher-agent/requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.5
prometheus-client>=0.16.0
